Issue #839 Move test suites to the validation folder
