+++ src/validation/search/__tests__/validation.tests.ts
@@ -1,5 +1,5 @@
 // tslint:disable:no-expression-statement
-import { toValidSearchData, toGeoCoderLatLong } from '../../../validation/search';
+import { toValidSearchData, toGeoCoderLatLong } from '..';
 import { aString, aNumber } from '../../../application/__tests__/helpers/random_test_values';
 
 describe('Search response validation', () => {
